+++ src/app/login/page.tsx
@@ -1,15 +1,14 @@
 'use client';
 
-import { Box, Flex, Text } from '@chakra-ui/react';
+import { Box, Button, Flex, Text } from '@chakra-ui/react';
 import { v4 as uuidv4 } from 'uuid';
 import KakaoLoginBtnSVG from '@/public/assets/KakaoLoginBtn.svg';
 import NaverLoginBtnSVG from '@/public/assets/NaverLoginBtn.svg';
 import LogoSVG from '@/public/assets/logo.svg';
 import { OAUTH } from '@/src/app/constants/auth';
 import Footer from '../components/Footer/Footer';
-import { useEffect } from 'react';
-import { toaster } from '@/src/components/ui/toaster';
-import { useSearchParams } from 'next/navigation';
+import { LoginErrorToast } from '../components/Toast/LoginErrorToast';
+import { useRouter } from 'next/navigation';
 
 const handleKakaoLogin = () => {
   window.location.href = OAUTH.KAKAO.AUTH_URL;
@@ -22,18 +21,17 @@ const handleNaverLogin = () => {
 };
 
 const Login = () => {
-  const error = useSearchParams().get('error');
-  useEffect(() => {
-    if (error) {
-      toaster.create({
-        title: '로그인 실패',
-        description: '로그인에 실패했습니다.',
-        type: 'error',
-      });
-    }
-  }, [error]);
+  const router = useRouter();
   return (
     <Flex width="100%" height="100vh" flexDirection="column">
+      <Button
+        onClick={() =>
+          router.replace('/login?error=Failed_to_get_access_token')
+        }
+      >
+        에러 발생
+      </Button>
+      <LoginErrorToast />
       <Box width="1920px" padding="38px 320px" borderTop="2px solid #f2f3f5">
         <Flex>
           <LogoSVG />
